packages/integrations/synap-haystack/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "synap-haystack"
version = "0.1.0"
description = "Synap memory integration for Haystack"
requires-python = ">=3.9"
license = "Apache-2.0"
dependencies = ["maximem-synap>=0.2.0", "haystack-ai>=2.0"]

[project.optional-dependencies]
dev = ["pytest>=7.0", "pytest-asyncio>=0.21"]

[tool.setuptools.packages.find]
where = ["."]
include = ["synap_haystack*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
